+++ common/student.cpp
@@ -8,7 +8,7 @@ namespace entity {
 
 Student::Student(const QString &firstName, const QString &lastName,
                  const QString &email, const QString &indexNumber,
-                 uint8_t yearOfStudy, uint64_t id, const QString &rfid)
+                 uint8_t yearOfStudy, quint64 id, const QString &rfid)
     : Person(firstName, lastName, email, id, rfid), mIndexNumber(indexNumber),
       mYearOfStudy(yearOfStudy) {}
 
@@ -73,7 +73,7 @@ QSqlQuery Student::insertQuery(const QSqlDatabase &database,
     QSqlQuery query(database);
     query.prepare("INSERT INTO STUDENT (ID, INDEX_NUMBER, YEAR_OF_STUDY) "
                   "VALUES (:id, :index_number, :year_of_study)");
-    query.bindValue(":id", static_cast<quint64>(student.id()));
+    query.bindValue(":id", student.id());
     query.bindValue(":index_number", student.indexNumber());
     query.bindValue(":year_of_study", student.yearOfStudy());
     return query;
@@ -86,7 +86,7 @@ QSqlQuery Student::updateQuery(const QSqlDatabase &database,
                   "INDEX_NUMBER = :index_number, "
                   "YEAR_OF_STUDY = :year_of_study "
                   "WHERE ID = :id");
-    query.bindValue(":id", static_cast<quint64>(student.id()));
+    query.bindValue(":id", student.id());
     query.bindValue(":index_number", student.indexNumber());
     query.bindValue(":year_of_study", student.yearOfStudy());
     return query;
